verification/requirements.txt
flask
faiss-cpu
google-generativeai
langchain
langchain-google-genai
scikit-learn
numpy
pandas

